Melee decomp 9/11: FT root
Match improvements for the ft root slice: ftcoll, ftcpuattack, ftdynamics,
ftafterimage, ftparts, ftdevice, and the new ft_459A data TU.

- ftcoll: .text 97.71 -> 99.18, .sdata2 -> 100, 11 functions to 100%
- ftcpuattack: .text 98.41 -> 99.57, .sdata2 -> 100, 4 functions to 100%
- ftdynamics: .text 99.79 -> 99.95, 2 functions to 100%
- ftafterimage: .text 90.91 -> 92.74, .sdata -> 100, .data -> 100
- ftdevice: split the ft_80459A68 overlay array into its three real
  objects (paired symbols.txt region, byte-conserving)
- ft_459A.c: new source for the previously source-less costume data unit
- Cross-module (codegen-verified byte-identical): ftkirby.c call-site
  cast removal for the typed ftParts_800755E8 signature; pltrick.h/c
  pl_80038144 attack_instance parameter widened to int with explicit
  u16 masks matching retail

+++ src/melee/pl/pltrick.c
@@ -204,7 +204,7 @@ static void fn_80037F00(HSD_GObj** gp, Fighter* fp, ft_800898B4_t* ev,
 }
 
 void pl_80038144(HSD_GObj* attacker_gobj, HSD_GObj* victim_gobj, s32 x18d4_int,
-                 ft_800898B4_t* ev_data, u16 attack_instance, s32 arg5,
+                 ft_800898B4_t* ev_data, int attack_instance, s32 arg5,
                  s32 source_ply)
 {
     Fighter* attacker_fp2;
@@ -256,8 +256,8 @@ void pl_80038144(HSD_GObj* attacker_gobj, HSD_GObj* victim_gobj, s32 x18d4_int,
                     victim_fp->player_id, victim_fp->x221F_b4, ev.x2073);
     }
 
-    if (attack_instance == 0 ||
-        victim_fp->dmg.x18ec_instancehitby != attack_instance)
+    if ((u16) attack_instance == 0 ||
+        victim_fp->dmg.x18ec_instancehitby != (u16) attack_instance)
     {
         *(s32*) &victim_fp->dmg.x18d4 = ev.x2070_int;
         victim_fp->dmg.x18ec_instancehitby = attack_instance;
